Move migration to separate function
Remove image store instantiation from image service
