+++ daemon/images/image_prune.go
@@ -2,20 +2,11 @@ package images // import "github.com/docker/docker/daemon/images"
 
 import (
 	"context"
-	"fmt"
-	"sync/atomic"
-	"time"
 
-	"github.com/docker/distribution/reference"
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/filters"
-	timetypes "github.com/docker/docker/api/types/time"
 	"github.com/docker/docker/errdefs"
-	"github.com/docker/docker/image"
-	"github.com/docker/docker/layer"
-	digest "github.com/opencontainers/go-digest"
 	"github.com/pkg/errors"
-	"github.com/sirupsen/logrus"
 )
 
 var imagesAcceptedFilters = map[string]bool{
@@ -31,138 +22,142 @@ var errPruneRunning = errdefs.Conflict(errors.New("a prune operation is already 
 
 // ImagesPrune removes unused images
 func (i *ImageService) ImagesPrune(ctx context.Context, pruneFilters filters.Args) (*types.ImagesPruneReport, error) {
-	if !atomic.CompareAndSwapInt32(&i.pruneRunning, 0, 1) {
-		return nil, errPruneRunning
-	}
-	defer atomic.StoreInt32(&i.pruneRunning, 0)
-
-	// make sure that only accepted filters have been received
-	err := pruneFilters.Validate(imagesAcceptedFilters)
-	if err != nil {
-		return nil, err
-	}
-
-	rep := &types.ImagesPruneReport{}
-
-	danglingOnly := true
-	if pruneFilters.Contains("dangling") {
-		if pruneFilters.ExactMatch("dangling", "false") || pruneFilters.ExactMatch("dangling", "0") {
-			danglingOnly = false
-		} else if !pruneFilters.ExactMatch("dangling", "true") && !pruneFilters.ExactMatch("dangling", "1") {
-			return nil, invalidFilter{"dangling", pruneFilters.Get("dangling")}
-		}
-	}
-
-	until, err := getUntilFromPruneFilters(pruneFilters)
-	if err != nil {
-		return nil, err
-	}
-
-	var allImages map[image.ID]*image.Image
-	if danglingOnly {
-		allImages = i.imageStore.Heads()
-	} else {
-		allImages = i.imageStore.Map()
-	}
-
-	// Filter intermediary images and get their unique size
-	allLayers := make(map[layer.ChainID]layer.Layer)
-	for _, ls := range i.layerStores {
-		for k, v := range ls.Map() {
-			allLayers[k] = v
-		}
-	}
-	topImages := map[image.ID]*image.Image{}
-	for id, img := range allImages {
-		select {
-		case <-ctx.Done():
-			return nil, ctx.Err()
-		default:
-			dgst := digest.Digest(id)
-			if len(i.referenceStore.References(dgst)) == 0 && len(i.imageStore.Children(id)) != 0 {
-				continue
+	return nil, errdefs.NotImplemented(errors.New("prune not implemented"))
+	/*
+			if !atomic.CompareAndSwapInt32(&i.pruneRunning, 0, 1) {
+				return nil, errPruneRunning
 			}
-			if !until.IsZero() && img.Created.After(until) {
-				continue
-			}
-			if img.V1Image.Config != nil && !matchLabels(pruneFilters, img.V1Image.Config.Labels) {
-				continue
-			}
-			topImages[id] = img
-		}
-	}
+			defer atomic.StoreInt32(&i.pruneRunning, 0)
 
-	canceled := false
-deleteImagesLoop:
-	for id := range topImages {
-		select {
-		case <-ctx.Done():
-			// we still want to calculate freed size and return the data
-			canceled = true
-			break deleteImagesLoop
-		default:
-		}
+			// make sure that only accepted filters have been received
+			err := pruneFilters.Validate(imagesAcceptedFilters)
+			if err != nil {
+				return nil, err
+			}
 
-		deletedImages := []types.ImageDeleteResponseItem{}
-		refs := i.referenceStore.References(id.Digest())
-		if len(refs) > 0 {
-			shouldDelete := !danglingOnly
-			if !shouldDelete {
-				hasTag := false
-				for _, ref := range refs {
-					if _, ok := ref.(reference.NamedTagged); ok {
-						hasTag = true
-						break
+			rep := &types.ImagesPruneReport{}
+
+			danglingOnly := true
+			if pruneFilters.Contains("dangling") {
+				if pruneFilters.ExactMatch("dangling", "false") || pruneFilters.ExactMatch("dangling", "0") {
+					danglingOnly = false
+				} else if !pruneFilters.ExactMatch("dangling", "true") && !pruneFilters.ExactMatch("dangling", "1") {
+					return nil, invalidFilter{"dangling", pruneFilters.Get("dangling")}
+				}
+			}
+
+			until, err := getUntilFromPruneFilters(pruneFilters)
+			if err != nil {
+				return nil, err
+			}
+
+			var allImages map[image.ID]*image.Image
+			if danglingOnly {
+				allImages = i.imageStore.Heads()
+			} else {
+				allImages = i.imageStore.Map()
+			}
+
+			// Filter intermediary images and get their unique size
+			allLayers := make(map[layer.ChainID]layer.Layer)
+			for _, ls := range i.layerStores {
+				for k, v := range ls.Map() {
+					allLayers[k] = v
+				}
+			}
+			topImages := map[image.ID]*image.Image{}
+			for id, img := range allImages {
+				select {
+				case <-ctx.Done():
+					return nil, ctx.Err()
+				default:
+					dgst := digest.Digest(id)
+					if len(i.referenceStore.References(dgst)) == 0 && len(i.imageStore.Children(id)) != 0 {
+						continue
 					}
+					if !until.IsZero() && img.Created.After(until) {
+						continue
+					}
+					if img.V1Image.Config != nil && !matchLabels(pruneFilters, img.V1Image.Config.Labels) {
+						continue
+					}
+					topImages[id] = img
+				}
+			}
+
+			canceled := false
+		deleteImagesLoop:
+			for id := range topImages {
+				select {
+				case <-ctx.Done():
+					// we still want to calculate freed size and return the data
+					canceled = true
+					break deleteImagesLoop
+				default:
 				}
 
-				// Only delete if it's untagged (i.e. repo:<none>)
-				shouldDelete = !hasTag
-			}
+				deletedImages := []types.ImageDeleteResponseItem{}
+				refs := i.referenceStore.References(id.Digest())
+				if len(refs) > 0 {
+					shouldDelete := !danglingOnly
+					if !shouldDelete {
+						hasTag := false
+						for _, ref := range refs {
+							if _, ok := ref.(reference.NamedTagged); ok {
+								hasTag = true
+								break
+							}
+						}
 
-			if shouldDelete {
-				for _, ref := range refs {
-					imgDel, err := i.ImageDelete(ctx, ref.String(), false, true)
-					if imageDeleteFailed(ref.String(), err) {
+						// Only delete if it's untagged (i.e. repo:<none>)
+						shouldDelete = !hasTag
+					}
+
+					if shouldDelete {
+						for _, ref := range refs {
+							imgDel, err := i.ImageDelete(ctx, ref.String(), false, true)
+							if imageDeleteFailed(ref.String(), err) {
+								continue
+							}
+							deletedImages = append(deletedImages, imgDel...)
+						}
+					}
+				} else {
+					hex := id.Digest().Hex()
+					imgDel, err := i.ImageDelete(ctx, hex, false, true)
+					if imageDeleteFailed(hex, err) {
 						continue
 					}
 					deletedImages = append(deletedImages, imgDel...)
 				}
-			}
-		} else {
-			hex := id.Digest().Hex()
-			imgDel, err := i.ImageDelete(ctx, hex, false, true)
-			if imageDeleteFailed(hex, err) {
-				continue
-			}
-			deletedImages = append(deletedImages, imgDel...)
-		}
 
-		rep.ImagesDeleted = append(rep.ImagesDeleted, deletedImages...)
-	}
+				rep.ImagesDeleted = append(rep.ImagesDeleted, deletedImages...)
+			}
 
-	// Compute how much space was freed
-	for _, d := range rep.ImagesDeleted {
-		if d.Deleted != "" {
-			chid := layer.ChainID(d.Deleted)
-			if l, ok := allLayers[chid]; ok {
-				diffSize, err := l.DiffSize()
-				if err != nil {
-					logrus.Warnf("failed to get layer %s size: %v", chid, err)
-					continue
+			// Compute how much space was freed
+			for _, d := range rep.ImagesDeleted {
+				if d.Deleted != "" {
+					chid := layer.ChainID(d.Deleted)
+					if l, ok := allLayers[chid]; ok {
+						diffSize, err := l.DiffSize()
+						if err != nil {
+							logrus.Warnf("failed to get layer %s size: %v", chid, err)
+							continue
+						}
+						rep.SpaceReclaimed += uint64(diffSize)
+					}
 				}
-				rep.SpaceReclaimed += uint64(diffSize)
 			}
-		}
-	}
 
-	if canceled {
-		logrus.Debugf("ImagesPrune operation cancelled: %#v", *rep)
-	}
+			if canceled {
+				logrus.Debugf("ImagesPrune operation cancelled: %#v", *rep)
+			}
 
-	return rep, nil
+			return rep, nil
+	*/
 }
 
+/*
 func imageDeleteFailed(ref string, err error) bool {
 	switch {
 	case err == nil:
@@ -209,3 +204,4 @@ func getUntilFromPruneFilters(pruneFilters filters.Args) (time.Time, error) {
 	until = time.Unix(seconds, nanoseconds)
 	return until, nil
 }
+*/
